coroutine: parallel_for_each: relax contraints
coroutine::parallel_for_each is a faster replacement for
seastar::parallel_for_each, but cannot be used in as many places
due to overtight constraints.

This patch relaxes the constraints:
 - The function is required to be invocable on iterator_traits::reference
   (or ranges::range_reference_t for the range variant) instead of
   value_type (ranges::range_value_t resp.). This allows calling functions
   that mutate the element.
 - The function itself is no longer required to be an rvalue. This is
   accomplished by making the function type parameter for the constructor
   a free parameter (so it can be 'lambda&&' while the class's Func parameter
   is 'lambda') and using class template deduction guides to link between
   the two in the same way that universal references work.

With this I can convert the many parallel_for_each() instances in
Scylla that happen to be in coroutines to coroutine::parallel_for_each().