coroutine: explain and mitigate the lambda coroutine fiasco
Due to the way lambda coroutines are defined and how Seastar
continuations work, it is easy to have a use-after-free with lambda
coroutines.

Fortunately, it is easy to mitigate.

This patch documents the problem, provides a lambda coroutine wrapper
and describes how to use it, and documents APIs that are safe for lambda
coroutines without the wrapper.

A unit test is included.