future: mark as [[coro_await_elidable]]
In clang 20, marking a coroutine type as [[coro_await_elidable]] allows
the compiler to merge the coroutine state blocks of nested coroutines:

    future<> c1(int& x) {
        int y = x + 3;
        co_await something(y);
    }

    future<> c2(int x) {
        co_await c1(x);
    }

The compiler now recognizes that the lifetime of c1 is fully contained
within that of c2, and stores all coroutine-local variables within
one allocation instead of two.

A test is added to verify the ellision.