net: posix_server_socket_impl: coroutinize accept()
The recursion is replaced by a while loop.

This simplifes later patches which add more optional awaiting
steps.

The coroutinization does not add an allocation, as typically accept()
will return a non-ready future and future::then_unpack() will allocate
a continuation. In any case accept() is not very performance sensitive.