net: posix_server_socket_impl: implement proxy protocol v2 port load balancing
Similar to port load balancing, but takes the port number from
a proxy protocol v2 header prepended to the connection data stream.

We can't use input_stream/data_source here since the connection will
be transferred to another shard, and these objects are deeply intertwined
with the shard, but we can use pollable_fd which is just a wrapper around
a file descriptor. We don't need to use the equivalent of read_exactly()
because the protocol guarantees the header will be sent in one packet
and not be fragmented.