net: posix_connected_socket: specialize for proxied connections
Proxied connections derive the local and remote addresses from
the proxy protocol, not getsockname(2)/getpeername(2). Add a new
derived class to hold the storage for those addresses and supply
new implementations for local_address() and remote_address().

A constructor in posix_connected_socket_impl is made public for this.
Since the class itself is part of the implementation, this doesn't
expose anything.